fix test_system initialize/finalize
Move calls to dlsym to test_system::ctor and add logic to assert they
are valid in the initialize function.

Implement test_system::finalize to clean up any internal structures
(like the fd map)

Modify munit-enum-c to call finalize in the TearDown function

+++ testing/unit/munit-enum-c.cpp
@@ -84,9 +84,12 @@ class enum_c_p : public ::testing::TestWithParam<const char *> {
   }
 
   virtual void TearDown() override {
-    std::string cmd = "rm -rf " + tmpsysfs;
-    std::system(cmd.c_str());
     EXPECT_EQ(fpgaDestroyProperties(&filter), FPGA_OK);
+    if (!tmpsysfs.empty() && tmpsysfs.size() > 1) {
+      std::string cmd = "rm -rf " + tmpsysfs;
+      std::system(cmd.c_str());
+    }
+    system_->finalize();
   }
 
   std::string tmpsysfs;
